UP-629 Anchoring V2 #comment Moving Tree Monitor stuff to tree package for better org.
