+++ event-log-chainer/src/main/scala/com/ubirch/chainer/services/tree/TreeCache.scala
@@ -1,4 +1,4 @@
-package com.ubirch.chainer.services
+package com.ubirch.chainer.services.tree
 
 import com.typesafe.config.Config
 import com.ubirch.chainer.models.Mode
@@ -16,7 +16,7 @@ class TreeCache @Inject() (@Named(MemCache.name) cache: Cache, config: Config)(i
   val LATEST_TREE: String = Mode.fold(mode)("LASTEST_SLAVE_TREE")("LATEST_MASTER_TREE")
 
   def prefix(value: String): String = {
-    val px = Mode.fold(mode)("sl.")(("ml."))
+    val px = Mode.fold(mode)("sl.")("ml.")
     if (!value.startsWith(px)) px + value else value
   }
