+++ event-log-chainer/src/main/scala/com/ubirch/chainer/services/tree/TreeEventLogCreator.scala
@@ -1,4 +1,4 @@
-package com.ubirch.chainer.services
+package com.ubirch.chainer.services.tree
 
 import com.typesafe.config.Config
 import com.typesafe.scalalogging.LazyLogging
